En teoria terminada la parte logica

+++ src/main/java/DAO/DAOBdInmueble.java
@@ -76,7 +76,7 @@ public class DAOBdInmueble implements InmuebleDAO{
     }
 
     @Override
-    public List<Inmueble> listAllByPropietario(int idPropietario) {
+    public List<Inmueble> listAllByPropietario(Long idPropietario) {
         EntityManager manager = Conexion.emf.createEntityManager();
         @SuppressWarnings("unchecked")
         List<Inmueble> lista = (List<Inmueble>) manager.createQuery("From Inmueble as i Where i.estado!='BAJA'").getResultList();
@@ -85,7 +85,7 @@ public class DAOBdInmueble implements InmuebleDAO{
     }
 
     @Override
-    public List<Inmueble> listAllByPropietario(int idPropietario, int inicio, int fin) {
+    public List<Inmueble> listAllByPropietario(Long idPropietario, int inicio, int fin) {
         EntityManager manager = Conexion.emf.createEntityManager();
         @SuppressWarnings("unchecked")
         List<Inmueble> lista = (List<Inmueble>) manager.createQuery("From Inmueble as i Where i.estado!='BAJA'").setMaxResults(fin - inicio + 1).setFirstResult(inicio - 1).getResultList();
